+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Refactor `Impl Future` parameters into `impl IntoFuture`. 
 * Implement `IntoFuture for ResponseVar<T>`.
+* Remove `.zr-apk` requirement of extension on the folder name.
 
 # 0.11.1