+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Refactor `.zr-apk` to not require to be inside the staging dir.
 * Refactor `Impl Future` parameters into `impl IntoFuture`. 
 * Implement `IntoFuture for ResponseVar<T>`.
 * Remove `.zr-apk` requirement of extension on the folder name.