+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+
+# 0.11.2
+
 * Implement initial `ColorScheme` for Android.
 * Support `RUSTFLAGS` "deny warnings" in cargo zng.
 * Warn when `.zr-copy` does not find the directory or file.